+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# Pending
+# Thu Nov  8 14:32:20 UTC 2018
+
+* Load square asset when constructing backup campaign [#7074](https://github.com/Applifier/unity-ads-webview/pull/7074)
+
+# Thu Nov  8 11:34:51 UTC 2018
 
 * Move trackingUrls to base Campaign model [#7053](https://github.com/Applifier/unity-ads-webview/pull/7053)
 * Fix OrganizationID Description [#6941](https://github.com/Applifier/unity-ads-webview/pull/6941)